requirements.txt
pandas
numpy
yfinance
sqlalchemy
psycopg2-binary
jupyter
matplotlib
scikit-learn
tensorflow
yahoo_fin
schedule